pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helix-agent"
version = "0.1.0"
description = "Python replication of the HelixAgent AI coding assistant"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "anyio>=4.0",
    "orjson>=3.9",
    "dataclasses-json>=0.6",
    "click>=8.1",
    "rich>=13.0",
    "prompt_toolkit>=3.0",
]

[project.scripts]
helix-agent = "helix_agent.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "mypy>=1.9",
    "ruff>=0.4",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["helix_agent*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
